Move all tests to test_endpoints
