+++ API/.env
@@ -2,5 +2,5 @@ source venv/bin/activate
 export FLASK_APP="run.py"
 export APP_MODE="dev"
 export SECRET="232adfadk32323jkadfo23i232323"
-export DATABASE_URL="postgresql://postgres:ed@localhost/book_a_meal"
+export SQLALCHEMY_DATABASE_URI="postgresql://postgres:ed@localhost/book_a_meal"
 export TEST_DATABASE_URL="postgresql://postgres:ed@localhost/book_a_meal_test"
